Move narrowphase to collision_functions.py
